requirements.txt
requests>=2.31,<3
pandas>=2.1,<3
pyyaml>=6.0,<7
schedule>=1.2,<2
pytest>=7.4,<9
